Add Ensembl checks to the tie-breaking code; avoid NAs
Previously, we were too overzealous with emitting NAs. Now, we don't
emit NAs if none of the genes appear in any Brainarray platform, and
instead we break ties by checking if any of the genes appear in the main
assembly on Ensembl. This method ends up breaking a lot of the ties.